[FIX] website_sale_account_taxcloud: update taxcloud tax before payment
### Setup
- DB with main company on the US
- account_taxcloud, ecommerce, sales, and website_sale_coupon installed
- Full address (US) set on the company
- Taxcloud activated (with test credentials*)
- Taxcloud fiscal position set as “detect automatically”
- Have a product available in your eShop

### Steps to reproduce
- in a new incognito window, go to the shop, put a product in your
  shopping cart and process to check out
- Fill in an Americana address. It's important to have an address that
  makes sense and whose state is supported by taxcloud. You can use the
  following:
  - City: Gaylord
  - Postcode: 55334
  - Country: United States
  - State: Minnesota
- click next, and you'll be redirected to a payment page
- now open the shopping cart in a new tab

If you go to the sales app and look at the quotation associated with
your session, you should see that the taxes have been removed.  However,
since you already have a payment page open, you can complete the
checkout without having to pay the taxes.

Cause:
In the `website_sale_coupon` module,
`SaleOrder.recompute_coupon_lines()` is called when the user goes to the
shopping cart.
However, `SaleOrder.recompute_coupon_lines()` removes the tax on
taxcloud lines. In most cases this is fine since the taxes are
recomputed when the user goes to the payment page.  But in this case,
the user already has a payment page open, so the tax is never
recomputed.

opw-2815397

closes odoo/enterprise#35241

X-original-commit: 328ef1b5fcfb80654f5724be5662566af1ecfbb6
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>