[FIX] website_sale_subscription: fix subscription product search in ecommerce
Steps to reproduce:

- Have ecommerce and subscription module installed.
- Create a subscription product.
- Go to the website and search for the subscription product.

Issue:

We are goint to get many tracebacks while trying to search for the
product, with a key error on `subscription_duration`. This is caused
because we never pass the right pricelist to
`_get_first_suitable_pricing_values` so we never get the
`combination_info` computed properly for subscription.

Solution:

Added the proper pricelist of the given product if we don't get any
other pricelist.

opw-3102681

closes odoo/enterprise#35250

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>