[FIX] stock_barcode: filter pickings using phone camera
On a list of pickings, it is currently not possible to scan a
product with the camera of the device (and therefore, the user can't
filter the pickings)

To reproduce the issue:
(Use demo data)
1. Create a product P:
   - Barcode: 123
2. Create and confirm a planned receipt R with 1 x P
3. Open the Barcode App
   - Note: There is a "Tap to Scan" button so a user can scan
     something thanks to the camera of the device
4. Operations > Receipts
   - There is a tip: "Scan a product to filter your records"

Error: According to the tip, it is possible to scan P to filter the
picking and get R. However, there isn't any button so the user can
scan a product with his camera.

Such a button is also present on a picking and an inventory adjustment.
So, considering this and the note of step 3, it seems it should
always be possible to scan with the camera of the device.

OPW-3110048

closes odoo/enterprise#35483

X-original-commit: 508266cb5776206f382aea53ba1c1468ecc79d1f
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>