[IMP] website_knowledge: adapt module to OWL
Since we are transitioning the TopBar as a widget, the element on which the
view was inheriting to add the button doesn't exist in the articles' form view
and we need to now patch the new TopBar in order to make this viable again.

This commit is adding a patch to the new TopBar Component in order for the
new button added by this module to be correctly placed and functional.

Part-of: odoo/enterprise#35261