[FIX] l10n_be_disallowed_expenses: translations to English and re-add lost in Transifex
Module was written in French while we want the code to be in English.
Also some translations were on Transifex on 15.0 and got lost in the transition to 16.0
since we have no Transifex project for them anymore.
This PR translates the module to English and reintroduce lost translations for French and Dutch.
We also remove i18n_extra that is both incorrect (wrong module name) and outdated.

closes odoo/enterprise#35268

Task: 3127437
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>