[IMP] l10n_pe_edi: Adding the XML structure to support ISC
ISC is the tax that you pay to the Sunat when you buy goods such as
alcoholic beverages, cigarettes or fuel. Support for this tax is
added to electronic invoices.

The Sunat requires that the "cbc:TaxableAmount" field reflect the amount
of the product base. Since there are two taxes on the same line, we must
subtract the value of the ISC that is used as the basis for calculating
the IGV.

Legal Reference:

Guide for the Preparation of Electronic Invoice:
https://cpe.sunat.gob.pe/sites/default/files/inline-files/gu%C3%ADa%20xml%20para%20factura%20electr%C3%B3nica%20%28ubl%202.0%29_0.pdf

Impuesto selectivo al consumo (ISC):
https://www.gob.pe/7918-impuesto-selectivo-al-consumo-isc

Task Latam: https://latam-localizations.odoo.com/web#id=927&menu_id=88&cids=1&action=188&model=project.task&view_type=form

closes odoo/enterprise#36722

X-original-commit: 62dbe831b0633eb6b40599bbe6075a59bb8deecf
Signed-off-by: Josse Colpaert <jco@odoo.com>