[FIX] knowledge: update the outline behavior when clicked on refresh
This commit fixes a bug where when you would click on the refresh button
on articles_structure and then access an article via one of the links,
the structure would be reset to its previous state.

This happened because the body of the article wasn't save in DB,
now we are checking when the props are patched and if there was a modification
and we save the record's data.

task-3113212

closes odoo/enterprise#35283

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>