[IMP] l10n_be_hr_contract_salary: improve fleet managment
Purpose: For now if the fleet manager wants to reserve a car for the
applicant that is not yet an employee, they reserve the car by linking it
to a fake employee "reserved" as a future driver. Afterwards, when the
manager sends the link, they force the car by putting it as an default
car, as otherwise the car will not appear in the list of the cars as the
future driver is set on it. It requires too much coordination.

Instead, it is already possible that the manager sets applicant
(corresponding res.partner) as a future driver on the car to reserve it.

In this commit, we compute the default car for the res.partner, where
they are set as the future driver, so that the fleet manager does not
have to manually look up for the vehicle and set it.

task - 3053243

closes odoo/enterprise#35907

Signed-off-by: Kevin Baptiste <kba@odoo.com>