[FIX] l10n_us_1099: only include active companies
lines_to_export is a stored computed field which by default has
compute_sudo=True. Because of this the multi-company record rules are
not applied.

This also stops manually calling _compute_lines_to_export in the test
and instead leaves it to the ORM so we can properly test this.

opw-3100845

closes odoo/enterprise#35403

X-original-commit: 0b9b75f1e77ee75d0bbb497112713bfc991dae84
Signed-off-by: William André (wan) <wan@odoo.com>