[FIX] knowledge: Sort trashed articles using write date
Purpose:
--------
In the trashed articles list, we currently order by default the articles
using their deletion date.
This does not work because the deletion_date field is computed and not
stored, and logs a Warning:
`Model 'knowledge.article' cannot be sorted on field 'deletion_date'
(not a column)`.

This commit sorts the articles in the trashed articles list using the
`write_date` field. This is the same as sorting them using the deletion
date, given the fact that the deletion date is simply the write date
plus a "trash delay" (which is the same for every article), except that
`write_date` is stored, and can therefore be used in a `order by`
statement.

Task-3114916

closes odoo/enterprise#35658

X-original-commit: f81e2c07fc5eb26120b88732835ac3684d230c66
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>