[FIX] delivery_ups: prevent sending "Bulk Content" as a barcode
Before this commit: if you don't put the product in a package, the package
name would be "Bulk Content". In case of sending the package inside the
US, we will send the package name to the UPS, but it has to be a barcode
and can not contain spaces.

The solution is first to check whether the package name contains spaces
or not.

opw-3054892

closes odoo/enterprise#36641

X-original-commit: 1bebc85d4d57d0452e16c06a5dbed0e69b04c860
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>