[FIX] delivery_ups: prevent sending "Bulk Content" as a barcode
Before this commit: if you don't put the product in a package, the package
name would be "Bulk Content". In case of sending the package inside the
US, we will send the package name to the UPS, but it has to be a barcode
and can not contain spaces.

The solution is first to check whether the package name contains spaces
or not.

opw-3054892

closes odoo/enterprise#36683

X-original-commit: c972594290a08a7de628d7b52da248fe0b47e904
Signed-off-by: William Henrotin (whe) <whe@odoo.com>