[FIX] helpdesk: fix the helpdesk dashbaord value
Before this commit, Helpdesk > Dashboard targets only accept
inputs with a dot as a decimal separator, even in languages
that use a comma and also it does not support the decimal places
values. For Ex: if user enters 25.56 it will only display the
25.00 so decimal was converted to interger by default.

So in this commit, check for the integer values and display only
interger values instead of float on dashbaord.

task-2942432

closes odoo/enterprise#35415

X-original-commit: b6389c5706ca2774795b174ee3d1166654f87227
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>