[FIX] knowledge: share panel will display if we go through trash
Current behavior before PR:

If the user opens the trashed article and then switches to any active article
via the tree panel, clicking on the share button share panel won't display.
This is because the state "displaySharePanel" wasn't updated for the above case.

Desired behavior after PR is merged:

The share button element added to dependency function of useEffect hook.
If there is a change in the button state, the useEffect will be re-executed and
expected behavior will work.

Task-3106375

closes odoo/enterprise#37592

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>