[FIX] helpdesk: fix tree view of closed tickets
Current behaviour:
Clicking an on an line in the tree view of the closed ticket analysis
doesn't do anything.

Expected behaviour:
Should redirect to the form view of the ticket that was clicked.

Steps to reproduce:
- Install Helpdesk
- Go to Helpdesk > Closed Tickets > click on a ticket
- Nothing happens ...

Reason for the problem:
The action that loads the closed tickets is not sending the form view
from the server.

Fix:
Add the form view in the action definition.

Affected versions:
- saas-15.2
- saas-15.3
- 16.0
- master

opw-3112442

closes odoo/enterprise#35457

X-original-commit: 14caa381e8303ae42acee9bed99fafff88046615
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>