[FIX] hr_payroll: fix test_dashboard_stats in start of the year
In payroll running the test_dashboard_stats test in the first month of the year
would cause the test to fail, as a contract would be created with a start date
in the last year, while for other months, any contract in the past was guaranteed
to be created in the current year. This commit fixes this by checking if the
first contract was created last year and adapting the amount of employees the
test expects for each year.

task-3116501

closes odoo/enterprise#35358

X-original-commit: 289bb716b842a41c59fd9f8ceaee40d9fd232369
Signed-off-by: Kevin Baptiste <kba@odoo.com>