[FIX] industry_fsm: remove duplicated project field in quick create
Before this commit, when the user wants to create a task in the kanban
view of `My Tasks` menu in project. The project field is in double in
the quick create.

This commit removes the duplicated field by creating a quick create form
view in primary used when the user selects a project in the main view of
Project app and wants to create a task into that project.

Part-of: odoo/enterprise#35377