[IMP] industry_fsm: avoid select internal project in task views
Before this commit, when the user creates a task in the (quick create)
form view of task, he can select the internal project. The goal of this
project is to contain the timesheets representing the time offs of each
employee.

This commit filter the projects available for the task creation/edition
to be sure the user cannot select internal project.

closes odoo/enterprise#35377

Related: odoo/odoo#109002
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>