[FIX] sale_renting: remove unsupported t-field in arch
This commit is the enterprise counter part of odoo/odoo#107083.
`t-field` isn't a valid qweb directive. As a matter of fact, the
commit introducing it noted that it wasn't working [1]. Now that
we validate qweb directives used in archs, this one would throw
an error. We thus remove it.

Task 3085357

[1] https://github.com/odoo/enterprise/commit/11a41bb72a1f045702f37eddc0f583a025112d0c#diff-97c66f033e6f370fadfb24d5cea5f6a09ab3e95ceda55b0c90e38aa1c9d24425R75

Part-of: odoo/enterprise#35352