[FIX] web_studio: entering X2M that contains a wowl Widget
Have a form view containing a one2many which in turn has a widget in its arch.
(from the registry "view_widgets", not a field widget).

Open studio, and navigate to edit that one2many list.

Before this commit, there was a crash as the legacy basic view tried to get some information
on the widget, which doesn't exist in the legacy registries.

After this commit, there is no crash. As we converted every legacy BasicView (form, list, kanban) to the new
WOWL framework, this legacy step is not necessary any longer, and obviously can cause crashes.

closes odoo/enterprise#35400

X-original-commit: 7163dc2d88e21438e2e44d31a5f64c24fb6bcb05
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>