[FIX] account_sepa: fix "Sepa to send" filter in dashboard
In accounting dashboard there is a link "XXX Payments to send via SEPA". It's
supposed to show list of SEPA payments, but instead it shows all payments for
the journal.

In v15 it works via context `search_default_sepa_to_send=1`, but that filter was
removed in v16 [1]

Fix it by forcing the domain on the action level.

[1]: https://github.com/odoo/enterprise/commit/1983b9949ac1d79536cd92bcece777a355acd113

opw-3101288

closes odoo/enterprise#38178

X-original-commit: 1b4f3ff0e82c27cc6bef542f2b8cd752492b7ec6
Signed-off-by: Quentin De Paoli <qdp@odoo.com>