[IMP] sale_planning: disable quick creation of role from product
Before this commit user can able to quick create a role from
product form view but creating a role without resource would
not be that useful and make not much sense.

So this commit disable quick creation of new roles from project
and make resources field required on `create and edit` form dialog
so SOL have at least a resource for role that was created from
product.

task-3117332

Part-of: odoo/enterprise#36540