[FIX] documents: fix "isUrlYoutube" compute method
Fixed the condition that is checking if a url is a YouTube url.

When the url is not a YouTube url, the "match" method returns null. So the condition in the return would consequently throw "TypeError: Cannot read properties of null (reading 'length')".

To reproduce the error, create a link document ("ADD A LINK"), put a non-youtube url. Then, click on another document of the page.

OPW-3113765

closes odoo/enterprise#35361

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>