[FIX] timesheet_grid: fix group_expand issue for project
Steps:
Create 80+ projects and a timesheet for each one.
Go to timesheets > all timesheets. Group by project.

Issue:
Some groups appear with no timesheet.

Cause:
The result is limited to 80 groups.
Then `_group_expand_project_ids` is called.
It adds the rest of it as empty groups to the result.

This method was used in timesheet reporting.
As only pivot and graph views are left in this action, it won't be
called anymore.

So, on one hand, it is not used anymore for its original purpose.
And on the other hand, it causes some issue for basic functionnalities.

Fix:
Deprecate it and return an empty set.

task-3084945

closes odoo/enterprise#35381

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>