[FIX] planning: template name considers working schedule
Current behaviour:
The planning templates names end time is not correct.

Expected behaviour:
The end time should be coherent with the working schedule.
A planning starting at 8AM with 10h duration should finish
at 10AM the next day and not at 6PM the same day.

Steps to reproduce:
- Install Planning
- Go to Configuration > Shift Templates
- Create a new template with start time at 8AM, duration 10h, Save
- Observe the template name is incorrectly named 8AM - 6PM

Reason for the problem:
Calculations for the end time of a planning wasn't taking into consideration
the working schedule of the role. It was a simple `start_time + duration`.

Fix:
Backport some code from 16.0 (there plannings takes work schedule into account
on the model). Fix the template name to use the `end_time` of the model, which is
correctly calculated, instead of the simple `start_time + duration`.

Affected versions:
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-3101105

closes odoo/enterprise#35526

X-original-commit: a3a19098c3619083a58c6b9eadb0f7a01ca6507c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>