[FIX] sign: use layout wizard not copying name into template
After https://github.com/odoo/enterprise/pull/22813, the copy method of
sign.template was overwritten to add a custom name to the template in case no
name was given in the copy call. As the use layout wizard uses the copy method
and wasn't passing a name explicitly, the name of the new template would be set
wrong. This commit fixes this by adding the name set in the attachment to
prevent the copy method of setting the wrong name.

task-3119880

closes odoo/enterprise#35418

Signed-off-by: Arnaud Joset <arj@odoo.com>