[FIX] appointment: prevent day overflow
There existed an issue when a user visited an appointment page on the website
at a time such that current_time + minimum lead time (min_schedule_hours)
was still in the same day but current_time + minimum lead time + apt duration
was in the next day (overflowing 24 hours).

In that case slots would be generated from that overflowed time
to the end hour of the current slot, on the next day.

Effectively, with appointments of 1 hour and a lead time of 12 hours
visiting these appointments at 11AM in the timezone of the appointment
would result in a number of time slots being duplicated.

Simply check that the date of the start datetime corresponds
to the day being processed.

task-3111413

closes odoo/enterprise#37079

X-original-commit: 45dd0884c84d0e8a31c2c36a975f1b1f7cf6d01c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>