[IMP] hr_payroll: add payslips to existing batch
Allows users to add payslips to existing batch. Previously, the user had
to create a new batch to incorporate new payslips.

Now when selecting the "Add Payslips" action, the users will be
presented with the list of payslips currently not in a batch to be
included in the current batch.

task-3103606

closes odoo/enterprise#35422

Signed-off-by: Kevin Baptiste <kba@odoo.com>