[IMP] stock_barcode: backorder dialog
With this commit, the validation of a picking from the Barcode App will
always by-pass the backorder's check. Instead, a new component,
`BackorderDialog`, will show which lines still waiting quantities.

The idea is to ignore the moves with no reserved quantities since they
are invisible in the Barcode App, and so, in case there are some moves
like that when the operation is validate, a backorder will automatically
be created (depending of the picking type's configuration) without
disrupting the picker's process.

For the move lines shown in the Barcode App, if they have reserved
quantity and their quantity done are lower than the reseved quantity,
they will be shown by the `BackorderDialog` for confirmation.
If some lines are uncompleted but other lines have no reserved quantity
for the same product, if those lines share the same `move_id`, their
`qty_done` will be summed. To be able to do that, when a new line is
created in the Barcode App, if it has the same product than the selected
line, the new line will copy the selected line's move.

task-3076044

Part-of: odoo/enterprise#35613