[IMP] sale_subscription: hide the new button in upsell
Purpose
======
The new button on the upsell screen is confusing for users as they think they are
creating an upsell when, in fact, they are just creating a new quote.

In this Commit, We hide the new button in the upsell screen so the user can't create a
manually.

task-3032430

closes odoo/enterprise#35445

Signed-off-by: Arnaud Joset <arj@odoo.com>