[FIX] helpdesk: fix message_new message timing
odoo/odoo#94018

The way tracking is handled is slightly modified to display
messages in the right chronological order.

A new message type is also added for tracking.

This commit applies some modifications to parts of the code
that relied on tracking being
handled during the creation process
and tracking tests that expected 'notification' message type.

Task-2834304

Part-of: odoo/enterprise#35466