[IMP] knowledge: ir.ui.view: bypass owl directives validation
The arch of the knowledge article form view is written as an owl
template. It uses several owl directives and a lot of js
implementation details, which is forbidden. PR odoo/odoo#107083
adds validation steps to detect forbidden owl directives in archs.
As a consequence, the knowledge form view would crash at validation.

This commit bypasses the validation for the knowledge form view.
This override could be removed as soon as that view is fixed (it
should become a client action).

Task 3085357

closes odoo/enterprise#35352

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>