[IMP] documents_account: import entry and bank statement actions
Add 2 document.workflow.action so that bank statement (e.g. CODA) and misc
operation  (e.g. SODA) can be imported from the document app. The behavior
is similar to the one found in the accounting dashboard.

Note: the no_document context is irrelevant in the account.move scope as it
does not inherit from 'documents.mixin'

task-id: 3116350
[enterprise](https://github.com/odoo/enterprise/pull/35486)

closes odoo/enterprise#35486

Signed-off-by: William André (wan) <wan@odoo.com>