[FIX] account_reports: add space between icon and selected filter
before this commit, in the accounting reports, the icon and selected filter in the date filter is showing without any spaces in between them.

after this commit, a space is added in between the calendar icon and selected filter as we have for the other filters in screen like partner filter and account filter.

closes odoo/enterprise#35499

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>