[IMP] l10n_co_edi: change value for GTE, GTA sections and EMI_4, ITT_3 tags
The values of the field `l10n_co_edi_fiscal_regimen` have been changed by the DIAN
several times in the past, the first version of UBL use the values with codes
48 and 49, but in the second version these codes didn’t match the new ones,
so we decided to use the code of the Tax Type as it was the same code for the
fiscal regimen.

In the Anexo 1.8 two new values were added in the `l10n_co_edi_fiscal_regimen`
that are not included in the Tax Type values.

So in this commit, I have changed EMI_4, ITT_3 tags and GTE, GTA section values
and added/updated the `l10n_co_edi_fiscal_regimen` selection in the partner.

TaskID - 3095064

closes odoo/enterprise#35555

X-original-commit: 9d64ccddc65da0566c6a92cb9f7748bb23bf7482
Signed-off-by: Josse Colpaert <jco@odoo.com>