[REF] various: cleanup usage of 'message_notify'
Notification method 'message_notify' is sometimes used as a standalone method
to send notifications. In that case it is called directly on MailThread
abstract model, notably when there is no context record or when it does not
inherit from 'mail.thread' directly. This is done mostly in technical models.

However in that case some model-specific code is not called, notably methods
computing default subject. In this commit we improve some calls to be sure
notifications have content enough to be understandable.

When possible, we also redirect the notification method on real records
inheriting from mail.thread, enabling a more complete process.

See community PR for more details.

Task-2710804 (Mail: Clean MailThread Posting API)

Part-of: odoo/enterprise#31017