[IMP] sale_subscription: adds close reason for auto closing
Currently it is difficult to analyse why subscriptions have been closed because
for most of them there is no close reason defined as OdooBot does not provide
any information when closing sub. This commit improves the close reasons and
implements a boolean field that prevents users from deleting close reasons that
are required in business flows.

task-3102584

closes odoo/enterprise#35536

Signed-off-by: Arnaud Joset <arj@odoo.com>