[FIX] account_3way_match: fix bills_to_pay filter and context
In some cases, for a purchase journal in the accounting dashboard, the
number of "Bills to Pay" will be different than the number of moves in
the list view generated by clicking on it.
There are two changes to make.

1) Change the `bills_to_pay` filter.
In the query that compute the bills to pay for the dashboard view, the
`payment_state` field is filtered as:
`payment_state in ('not_paid', 'partial')`
but in the filter `bills_to_pay`, we find:
`('payment_state', 'not in', ('paid', 'in_payment'))`
so an `account.move` with a `payment_state == 'reversed'` would be taken
into account by the filter but not by the query.
I've change the filter to match the query.

The query is defined in the function
`_get_open_bills_to_pay_query(self)` in
enterprise/account_3way_match/models/account_journal_dashboard.py

2) Change the context for the "Bills to Pay" link.
When clicking the `Bills to Pay` link, it opens an `account.move` list
view with 3 filters: `Vendor Bills` & `Not Paid` & `Bills to Pay`.
This module is responsible for the last 2 filters.  The `Not Paid` one
should be remove, as the `payment_state` is taken into account in the
`Bills to Pay` filter.
Originally (at the creation of the module), that was not the case and
that's why this filter was added.  Because the`payment_state` domain
is now different than `('payment_state', '=', 'not paid')`, this filter
clashes with the `Bills to Pay` filter and so partially paid moves will
not be shown.

opw-3103219

closes odoo/enterprise#35865

X-original-commit: 32421fa933a5f0aee0d9e5927f242fb5bb9f99c9
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Arnaud-Sibille <arsi@odoo.com>