[FIX] account_reports: User proper date filter when auditing Aged Payable/Receivable's journal items
To reproduce, on a db with only the demo data:

1) Open the Aged Payable/Receivable.
2) Select a report date 2 months in the future.
3) Unfold Azure Interior's line. It has one late invoice, from the current month.
4) Click on "Journal Items" for Azure Interior. An empty tree view is shown.

===> This is because the "Report Dates" filter uses a date_from and a date_to. Despite the fact Aged Payable/Receivable both only allow the selection of one date, implementation reasons make it so that their options dict contains both such dates : so the Report Dates filter here only applies to this range, which is not what we want.

To select all the entries before the date_to, we add a new filter in the community counterpart of this branch. Existing users will have to update their database to benefit from this fix. Without doing so, the report will consider all the move lines and ignore the report dates.

closes odoo/enterprise#35588

Related: odoo/odoo#109457
Signed-off-by: Laurent Smet <las@odoo.com>