[FIX] sale_renting: unsupported datetime format
The hardcoded format to send the date values to open the wizard
is causing tracebacks:

* when editing a rental line
* when adding a second rental line in an order

in some specific languages.

This commit makes sure the format of the datetime sent (in the context)
to the server is supported.

Also makes sure non rental lines do not impact the default
start/end date given as wizard default values.

opw-3089674

closes odoo/enterprise#35591

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>