[FIX] sign: fix getting sign values from elements
Getting the values from each sign item in sign was done in general way,
meaning that it would try to get the item's `.val()` and then check for special
cases like selection, checkbox. This was confusing and caused some bugs like
https://github.com/odoo/enterprise/pull/31227, opw-3045307 and opw-3101910.
This commit explicitly handles each case, making it easier to customize
behavior for each type of sign item.

task-3109473

closes odoo/enterprise#35682

Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>