[FIX] knowledge: save when copying article
Before this commit when you tried to create a private copy of an article
the changes that were made to the said article were not saved beforehand.
This means that the changes were not added to the private copy.

Now we are saving the article if changes were made before copying it.

We also added a new method, `_saveIfDirty`, that saves the record if it has
been dirtied by a modification and refactored the file to use it.

task-3126137

Correction Proposition

closes odoo/enterprise#35608

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>