[FIX] documents_spreadsheet: allow insert for users
Since changes made here: odoo/enterprise/pull/35560 the access rights had
been removed for the document users and set only for the Settings user.

In a recent fix: odoo/enterprise/pull/35560 the access rights have been
changed to include the documents manager.

But what we want is to allow the document users, so we adapted it to
allow the users.

closes odoo/enterprise#35650

X-original-commit: 95c9e575dba71044c6d25936167f653a423cc135
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>