[FIX] account_invoice_extract,sale_subscription: prevent race condition
Before this commit, when an invoices are posted, the ocr cron is launched.
This cron will write on the account.move records.

As subscription creates invoices in a cron and sends them once posted, race
conditions occurs frequently. The invoice is posted and then sent. Write occurs
on the account.move records after post. These two cron are trying to update the
records at the same time and, depending on the conditions, some write can occur
at the same time, causing race condition.

task id: 3054994

closes odoo/enterprise#35710

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>