[FIX] timesheet_grid: read worked hours with multi-company rule disabled
This revision is a follow-up of
odoo/enterprise#33526
odoo/enterprise#34372

When disabling the employee multi-company record rule,
the fix done in odoo/enterprise#33526 triggers the below bad SQL query:
```
bad query:
            SELECT aal.employee_id as employee_id, COALESCE(SUM(aal.unit_amount), 0) as worked_hours
            FROM account_analytic_line aal
            WHERE aal.employee_id IN () AND date >= '2021-04-01' AND date <= '2021-04-30'
            GROUP BY aal.employee_id

ERROR: syntax error at or near ")"
LINE 4:             WHERE aal.employee_id IN () AND date >= '2021-04...
```

Because
```py
self.env.cr.execute(query, (tuple(employees.ids), date_start, date_stop))
```

receives an empty list of ids `employees.ids`
because `employees` are forced within the code to belong to the user
companies, despite the record rule being disabled.
```py
employees = self.env['hr.employee'].with_context(active_test=False).search([
    ('id', 'in', [data['id'] for data in employees_grid_data]),
    ('company_id', 'in', self.env.companies.ids),
])
```

Letting users disabling multi-company rule is tolerated,
and so, the code should be resilient to that fact.
Hence, this piece of code shouldn't hard-code to filter out
employees which do not belong to the current user companies.

odoo/enterprise#33526 solves an access right issue
when the grid view is trying to display a timesheet in the current user
company but assigned to an employee from another company.
It did so by not displaying at all the timesheets in the grid view
assigned to employees not part of the user companies,
even if the timesheet itself was in the user companies.

The above behavior, while it does solve the reported issue for that
revision, is not perfect as it no longer displays in the grid view
timesheets you are entitled to see, as they are in your current
companies. But, because they are assigned to employees of other companies,
it removes them from the grid view.
For instance, you can see these timesheets assigned to employees of other
companies in the tree view.
Why couldn't you see in the grid view timesheets you can see in the
tree view ?
Another case is an employee changing of company:
Let's say an employee is employed Odoo Belgium,
log some timesheets while working in Odoo Belgium,
than move to Odoo US.
Suddenly, his manager in Odoo Belgium will no longer see in the grid
view his timesheets, which is incorrect.

Instead of not allowing the user to see those kind of timesheets
in the grid view, just read the data the user lacks the access right to
as sudo (`employees.resource_id`).
That way, the user still sees the timesheets in the grid view,
even for employees of other companies.

This revision adds unit test for the 3 cases:
- The case odoo/enterprise#33526 solves,
- The case odoo/enterprise#34372 solves,
- This new case of the disabling the multi-company rule.

closes odoo/enterprise#35719

X-original-commit: b46e98768de29ed1608c8dcf3ce75250ea49e092
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>