[IMP] spreadsheet_edition: zip images to xlsx file
This commits implements the zipping of images when exporting xlsx file.

When exporting an o-spreadsheet to excel file, the files are sent by
the client to be zipped by the server (we don't have any zip lib
client-side)

To reduce the networking load, only the image path is sent.
The server then replaces the path with the image data in the xlsx file.

task 3125701

closes odoo/enterprise#38296

Related: odoo/odoo#115514
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>