[FIX] account_inter_company_rules: add missing invoice_date_due
How to replicate the issue ?

1. Setup a company A and a company B
2. Enable invoice & bill synchronisation between the two
3. Create an invoice from company A with company B as the custommer
4. Fill the form and set a due date to somewhere in the future

If we know go to Accounting > Vendors > Bills on the company B side. We
can see that the due date is always set to the invoice date.

This is because of this line:
https://github.com/odoo/odoo/blob/ace565ee49d126b53d23cc97ae4cdea2626a9ea9/addons/account/models/account_move.py#L970

Because self.invoice_date_due is always set to False, the method will
always return self.invoice_date

opw-3111669

closes odoo/enterprise#35861

X-original-commit: 4368e5e23ed56e82a5874cca548f17f2b8cdd187
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>