[FIX] sale_temporal: pricing view from quotation template
When clicking on the button "prices" for a product from a
subscription plan (quotation template), the user lands on
the list of time based price rules that are compatible with
the recurrence of the quotation template.
The problem is that from there it is possible to create and
edit rules for other products (even non recurring one), other
recurrence, ... which is confusing and error prone.

New behaviour :
From that list it should only be possible:
- Edit the price or the pricelist of an existing rule 
- Create new rules for the "selected product" (product.product)
  and the "recurrence" of the template (these should be readonly)

This PR also move SO template field related to sale_temporal
from sale_subscription to sale_temporal.

task-id : 3076793

closes odoo/enterprise#35463

Signed-off-by: Arnaud Joset <arj@odoo.com>