[FIX] social: bad usage of fragment_to_query_string
Post an image on instagram via Odoo, traceback in the logs while serving
the callback. A `str` cannot be interpreted as a `Response`.

The `@route` decorator should be the first decorator (the one at the
top) used to decorate an endpoint. All other decorators should be placed
in between `@route` and the endpoint.

See also: odoo/odoo#109161

closes odoo/enterprise#35680

Signed-off-by: Julien Castiaux <juc@odoo.com>