[FIX] website_event_track_gantt: owl gantt view conversion
The js_class `event_track_gantt` was removed with commit 41e1316
As it do not exist anymore we should remove the js_class attribute
and adapt the `gantt-popover` template.

Before this commit, a traceback would occur as the unexisting js_class
falls back to the new Owl Gantt View implementation.

Part-of: odoo/enterprise#36202
Co-authored-by: Bruno Boi <boi@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>