[FIX] website_sale_renting: check if the product is rental
Steps to reproduce:
- install "website_sale" and "sale_renting" modules;
- for a product, add an accessory product which can be rented;
- go to ecommerce;
- add the product to the cart;
- view the cart.

Issue:
A traceback appears.

Cause:
We are trying to access a dictionary value with
a key which does not exist and more particularly,
we try to access the value of a product that can be rented.

Solution:
Test if a product can be rented using the `is_rental` key
before accessing its information.

opw-3117535

closes odoo/enterprise#35695

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>