[FIX] account_general_ledger: tax report options
The general ledger uses the generic tax report to generate some lines for sales and purchases journals.

The problem was that if your company was in a country that had a variant of the tax report, it will try to use that report instead. That would create an inconsistency between report ids that would raise an error.

This commit forces the use of the generic tax report for this case, keeping the report ids consistent.

closes odoo/enterprise#35700

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>