[FIX] web_enterprise: web_enterprise isn't autoinstall anymore
Since the merge of PWA in enterprise the dependency of `web_enterprise`
has changed (`base_setup` was added) and so the `auto_install` rules
isn't satisfied anymore.

This commit forces to auto install `web_enterprise` when web is
installed also when `base_setup` is not.

PS: another fix will be to set `auto_install: True` in `base_setup`.

Ref
[1] odoo/enterprise#35322

closes odoo/enterprise#35712

Signed-off-by: Raphael Collet <rco@odoo.com>