[FIX] {|l10n_be}hr_contract_salary: fix sim. link _compute_from_job
The fix from the task-3106778 introduced an issue where
it was not possible to compute the correct contract_id
when the job of the employee is changed.
This commit fixes the issue.

This commit also fixes an issue in l10n_be_hr_contract_salary
where the generate_simulation_link_action button attribute changes
allows to see the button on non-belgian company contract templates
while it shouldn't.

closes odoo/enterprise#35720

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>